Split out test token removal
